refactor resolver methods to service methods, add subscription to monitor progress
