+++ api/src/common/subscription-token.ts
@@ -1,4 +1,5 @@
 export class SubscriptionToken {
   static TranslationReport = 'TranslationReport';
   static ZipMapReport = 'ZipMapReport';
+  static DataGenerationReport = 'DataGenerationReport';
 }
